Move ObjectMapper instantiation into Provider
