+++ graylog2-shared/src/main/java/org/graylog2/shared/bindings/providers/ObjectMapperProvider.java
@@ -1,0 +1,53 @@
+/**
+ * This file is part of Graylog.
+ *
+ * Graylog is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Graylog is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.graylog2.shared.bindings.providers;
+
+import com.codahale.metrics.json.MetricsModule;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.databind.module.SimpleModule;
+import com.fasterxml.jackson.datatype.guava.GuavaModule;
+import com.fasterxml.jackson.datatype.joda.JodaModule;
+import org.graylog2.shared.jackson.SizeSerializer;
+import org.graylog2.shared.rest.RangeJsonSerializer;
+
+import javax.inject.Provider;
+import javax.inject.Singleton;
+import java.util.concurrent.TimeUnit;
+
+@Singleton
+public class ObjectMapperProvider implements Provider<ObjectMapper> {
+    protected final ObjectMapper objectMapper;
+
+    public ObjectMapperProvider() {
+        this.objectMapper = new ObjectMapper()
+                .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
+                .setPropertyNamingStrategy(new PropertyNamingStrategy.LowerCaseWithUnderscoresStrategy())
+                .registerModule(new JodaModule())
+                .registerModule(new GuavaModule())
+                .registerModule(new MetricsModule(TimeUnit.SECONDS, TimeUnit.SECONDS, false))
+                .registerModule(new SimpleModule()
+                        .addSerializer(new RangeJsonSerializer())
+                        .addSerializer(new SizeSerializer()));
+    }
+
+    @Override
+    public ObjectMapper get() {
+        return objectMapper;
+    }
+}
